requirements.txt
# Created: 2026-02-25 00:00
fastapi
uvicorn[standard]
pillow
numpy
scipy
torch
torchvision
